requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.1.1
idna==3.10
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
schedule==1.2.2
SQLAlchemy==2.0.35
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.2.3